+++ noodles-util/src/variant/io/reader.rs
@@ -26,7 +26,7 @@ where
     /// use noodles_vcf as vcf;
     /// use noodles_util::variant;
     ///
-    /// let data = Cursor::new(b"##fileformat=VCFv4.4
+    /// let data = Cursor::new(b"##fileformat=VCFv4.5
     /// #CHROM\tPOS\tID\tREF\tALT\tQUAL\tFILTER\tINFO
     /// sq0\t1\t.\tA\t.\t.\tPASS\t.
     /// ");
@@ -52,7 +52,7 @@ where
     /// use noodles_vcf as vcf;
     /// use noodles_util::variant;
     ///
-    /// let data = Cursor::new(b"##fileformat=VCFv4.3
+    /// let data = Cursor::new(b"##fileformat=VCFv4.5
     /// #CHROM\tPOS\tID\tREF\tALT\tQUAL\tFILTER\tINFO
     /// sq0\t1\t.\tA\t.\t.\tPASS\t.
     /// ");